move test/utils.js to where it belongs (#23586)
