+++ tests/helpers/diff-files.js
